requirements-dev.txt
--requirement requirements.txt

ipdb
ipython

django-debug-toolbar

mock
